Updated JLLVM, adding a bunch of features, including execution-engine and target-data support.

+++ src/jllvm/llvm/TargetJNI.java
@@ -11,7 +11,7 @@ package jllvm.llvm;
 class TargetJNI {
   public final static native void LLVMInitializeAllTargetInfos();
   public final static native void LLVMInitializeAllTargets();
-  public final static native long LLVMInitializeNativeTarget();
+  public final static native int LLVMInitializeNativeTarget();
   public final static native long LLVMCreateTargetData(String jarg1);
   public final static native void LLVMAddTargetData(long jarg1, long jarg2);
   public final static native String LLVMCopyStringRepOfTargetData(long jarg1);
